revision: invert class dependencies
